- Added new EB_GRID_EDITOR_ITEM, with which editor style texts can be clicked in and selected. - Equipped the Error and Warning tool with the new item. - Made the Error and Warning tool single item selectable. - Enable/Disable Cut/Copy contextual menu items according to selection status. - Initialize `y_offset' in EB_EDITOR_TOKEN_TEXT to fix the bug that selecting shifts the text by one pixel. - Added protections in the editors against calls on void when the editor does not depend on dev window. - In EB_EDITOR, use `font' and `line_height' from EDITABLE_TEXT_PANEL rather than SHARED_EDITOR_FONT to inherit the customizing abilities. - Fixed the problem in {CLICKABLE_TEXT}.end_processing that when there was only one line to display, coloring didn't process.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@76083 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/Eiffel/interface/new_graphical/text_window/interface/eb_custom_widgetted_editor.e
@@ -624,12 +624,16 @@ feature {NONE} -- Implementation
 			if mouse_move_idle_timer /= Void and then not mouse_move_idle_timer.is_destroyed then
 				mouse_move_idle_timer.destroy
 			end
-			dev_window.window.focus_in_actions.prune_all (check_search_bar_visible_procedure)
+			if dev_window /= Void then
+				dev_window.window.focus_in_actions.prune_all (check_search_bar_visible_procedure)
+			end
 			if search_tool /= Void then
 				search_tool.first_result_reached_actions.prune_all (first_result_reached_action)
 				search_tool.bottom_reached_actions.prune_all (bottom_reached_action)
 			end
-			search_bar.destroy
+			if search_bar /= Void then
+				search_bar.destroy
+			end
 			if customizable_commands /= Void then
 				customizable_commands.wipe_out
 			end
@@ -761,9 +765,9 @@ feature {NONE} -- Internal implentation cache
 			-- Note: Do not use directly!
 
 ;indexing
-	copyright:	"Copyright (c) 1984-2008, Eiffel Software"
-	license:	"GPL version 2 (see http://www.eiffel.com/licensing/gpl.txt)"
-	licensing_options:	"http://www.eiffel.com/licensing"
+	copyright: "Copyright (c) 1984-2008, Eiffel Software"
+	license:   "GPL version 2 (see http://www.eiffel.com/licensing/gpl.txt)"
+	licensing_options: "http://www.eiffel.com/licensing"
 	copying: "[
 			This file is part of Eiffel Software's Eiffel Development Environment.
 			
@@ -774,19 +778,19 @@ feature {NONE} -- Internal implentation cache
 			(available at the URL listed under "license" above).
 			
 			Eiffel Software's Eiffel Development Environment is
-			distributed in the hope that it will be useful,	but
+			distributed in the hope that it will be useful, but
 			WITHOUT ANY WARRANTY; without even the implied warranty
 			of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-			See the	GNU General Public License for more details.
+			See the GNU General Public License for more details.
 			
 			You should have received a copy of the GNU General Public
 			License along with Eiffel Software's Eiffel Development
 			Environment; if not, write to the Free Software Foundation,
-			Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
+			Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
 		]"
 	source: "[
 			 Eiffel Software
-			 356 Storke Road, Goleta, CA 93117 USA
+			 5949 Hollister Ave., Goleta, CA 93117 USA
 			 Telephone 805-685-1006, Fax 805-685-6869
 			 Website http://www.eiffel.com
 			 Customer support http://support.eiffel.com
